added generic convenience fns
