+++ pkg/builder/resource/spec_types.go
@@ -1,0 +1,7 @@
+package resource
+
+// SpecResource defines required methods for implementing a resource with spec.
+type SpecResource interface {
+	// CopyTo copies the content of the spec to a parent resource.
+	CopyTo(parent ObjectWithSpec)
+}
